update on several tables
Entity describes a "Row" of the database, so the table name is singular.  (Sale, not Sales.  Donation, not Donations).
Donation table references "Person" table.    Does Person already have "email" in it?  It should if it doesn't."
Sale table.   Quantity is whole (counting) numbers.  So what data type should we use for "Quantity"?